Creación de algunos repositorios JPA y la vista de clientes

+++ src/main/java/sidic/entities/Articulo.java
@@ -59,7 +59,7 @@ public class Articulo implements Serializable {
     @Id
     @Basic(optional = false)
     @Column(nullable = false, precision = 22)
-    private Double codigo;
+    private Long codigo;
     @Column(length = 35)
     private String referencia;
     @Column(length = 15)
@@ -116,15 +116,15 @@ public class Articulo implements Serializable {
     public Articulo() {
     }
 
-    public Articulo(Double codigo) {
+    public Articulo(Long codigo) {
         this.codigo = codigo;
     }
 
-    public Double getCodigo() {
+    public Long getCodigo() {
         return codigo;
     }
 
-    public void setCodigo(Double codigo) {
+    public void setCodigo(Long codigo) {
         this.codigo = codigo;
     }
